chore: update dependencies and improve settings/system integrations
- Update deps: notify 8.0, portable-pty 0.9, resvg/usvg 0.46, rusqlite 0.38, sysinfo 0.37
- Remove pulldown-cmark dependency (unused)
- Settings action now opens config.ts in user's configured editor with folder context
- Fix rusqlite 0.38 compatibility (i64 params for LIMIT/OFFSET, COUNT return type)
- Remove default notes hotkey - users must explicitly configure one
- Rewrite system preferences to use URL scheme for macOS Ventura+ compatibility

+++ src/clipboard_history/database.rs
@@ -449,7 +449,7 @@ pub fn get_clipboard_history_page(limit: usize, offset: usize) -> Vec<ClipboardE
     };
 
     let entries = stmt
-        .query_map(params![limit, offset], |row| {
+        .query_map(params![limit as i64, offset as i64], |row| {
             Ok(ClipboardEntry {
                 id: row.get(0)?,
                 content: row.get(1)?,
@@ -542,7 +542,7 @@ pub fn get_clipboard_history_meta(limit: usize, offset: usize) -> Vec<ClipboardE
     };
 
     let entries = stmt
-        .query_map(params![limit, offset], |row| {
+        .query_map(params![limit as i64, offset as i64], |row| {
             Ok(ClipboardEntryMeta {
                 id: row.get(0)?,
                 content_type: ContentType::from_str(&row.get::<_, String>(1)?),
